refactor: combine findable and registered event handler lambdas in update-doi-handler #NP-45622
