+++ datacite-update-doi-handler/src/test/resources/doi_publication_event_empty_customer_id_and_publication_id.json
