+++ settings.gradle
@@ -2,7 +2,7 @@ rootProject.name = 'nva-doi-registrar-client'
 include 'assign-doi-datacite'
 include 'doi-commons'
 include 'datacite-draft-doi-handler'
-include 'datacite-findable-doi-handler'
+include 'datacite-update-doi-handler'
 include 'datacite-delete-draft-doi-handler'
 include 'datacite-jaxb'
 include 'datacite-registered-doi-handler'
